MB-67250: Fix final mc_time narrowing
Add comments and narrow_cast around places that reduce the
64-bit time values to 32-bit time values. Avoding use of
gsl::narrow (throwing cast) in this core code for now.

Future plans for MB-67776 could make the casts safe, but
for now accept that uptime reaching 136 years is unlikely
and not worth more churn and testing.

Reviewed-on: https://review.couchbase.org/c/kv_engine/+/231204
Tested-by: Build Bot <build@couchbase.com>

+++ programs/engine_testapp/mock_server.cc
@@ -129,7 +129,7 @@ static uint32_t mock_limit_expiry_time(uint32_t t, std::chrono::seconds limit) {
     auto upperbound = mock_abstime(mock_get_current_time()) + limit.count();
 
     if (t == 0 || t > upperbound) {
-        t = upperbound;
+        t = gsl::narrow<uint32_t>(upperbound);
     }
 
     return t;
